requirements.txt
streamlit
google-genai
pypdf
faiss-cpu
python-dotenv
tiktoken
numpy
pdf2image